split subparser tests
