+++ .gitignore
@@ -14,3 +14,4 @@ dist/
 site/
 
 .~lock.*
+avatar.*
